fix: resolve Coverity medium issues
- SWAPPED_ARGUMENTS ×4 (sp2md.hpp): use named locals B_op_m/B_op_n to clarify transposed CSR dims
- UNINIT_CTOR ×2 (itsol_options.hpp, sypr.hpp): add in-class initializers for setby and e_offset
- UNINTENDED_INTEGER_DIVISION ×1 (convert.cpp): cast nnz to double before dividing by m

AMD-Internal: CPUPL-7944

+++ library/src/conversion/aoclsparse_convert.cpp
@@ -64,7 +64,7 @@ aoclsparse_int aoclsparse_opt_blksize(aoclsparse_int        m,
     double pc_blks_inc[2];
     double pc_diff_blks    = 0;
     double pc_diff_blkutil = 0;
-    double nnzpr           = nnz / m;
+    double nnzpr           = static_cast<double>(nnz) / m;
 
     for(int i = 0; i < 3; i++)
     {
